Fix flaky tests
Using date constructs using current timestamp in mocks can lead to
flaky tests. In this case, the mock dates are evaluated when the mocks
are imported, while the dates that are comparing to them are evaluated
when the tests run. This leads to failed runs when the mocks are
imported, the clock reaches the next full minute and then the test
runs.

The solution is to either hardcode the dates (like I've done) or use
some test tool to freeze time (but that's harder here since it would
mean a change in architecture).