fix(application-templates-health-insurance): Patch flaky test (#4091)
This test doesn't seem to be testing what the tester thinks it is.
The `oldestDraftId` will always be the last item created due to all
created attributes being the same, unless a clock tick occurs between
initializing the first external application and the second one - then it
will be the first external application, which for some reason has the
same UUID as the "current" application.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>