fix(island-ui): Wonky `AccordionItem` state props behavior (#5024)
* fix: Change prop signature to better reflect the internal behavior

This is still a bit weird because of how `onToggle` is tight-coupled
with togging the local-state, but whatever...

It's also unclean how `onClick` is meaningfully different from
`onToggle`, but again whatever...

BTW on the controlled variant, then `expanded` SHOULD be required

* style: Stop export of private types, add new component-specific types

* style: Rename `handleToggle` function to reflect its actual behavior

* fix: Fix broken `startExpanded` behavior (closes #5021)

The dependency array on the `useEffect` hook triggered an aggressive
reset every time `startExpanded` changed.

Also, if `startExpanded` and `expandedProp` were both set and `true`,
then the component started in a **closed** state.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>