chore(ci): Skip application-system-form e2e tests
Since they're super flaky, not providing value at the moment and not much progress in adding this type of e2e tests.

Need to have less flaky CI for this release cycle. Should enable again when we start writing this type of e2e tests and we have a way to fix the CI issue.