chore(ci): Skip flaky application-system-form-e2e test (#5855)
Since they're super flaky, not providing value at the moment and not much progress in adding this type of e2e tests.

Need to have less flaky CI for this release cycle. Should enable again when we start writing this type of e2e tests and we have a way to fix the CI issue.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>