feat(application-templates-driving-license): Changes to the payment field (#6117)
* feat(driving-license): Payment & consistency fixes

There's a lot of problems with the current implementation of payments in
the application system. This is probably step 1 - to allow multiple
types of payments in the same application.

Also better error handling and management when it comes to payments and
application state.

All in all the following changes are addressed here:

- Multiple types of cleanup in the application payment controller
- Allow more than one payment type per application
-

* think fix

* last bit of changes with testing

* upd

* should address the test and support the v1 version

* adding test for the success/error handler

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>