refactor(testing): Remove generic typing from testServer (#6577)
* refactor(testing): Remove generic typing from testServer

The generic typing in testServer is unneccesary.
The function itself does not use the type for anything than to pass it along to InfraModule
Changing the typing in the options object allows TypeScript to infer the type.

The generic version was causing typescript errors on usage even though it built fine.

Also changes to the useDatabase hook since Sequelize 5 and lower use
Bluebird's library version of promises, while hooks are expected to return ES6 promises.

Simple solution here is to wrap the Bluebird promise in a ES6 one.
A better solution might be some sort of typescript override.

* refactor(testing): Simplify the Bluebird promise issue with await

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>