fix(as): Fix slider on mobile devices (#6601)
Was using old-school touch events incorrectly. Refactored to use pointer
with one code to consistently support all pointing devices.

The Slider component and useDrag is copied in a few places. Slider has
additionally been modified here and there unfortunately and there is a
rush to get this into the release so I didn't have a chance to refactor
further.

Hopefully someone gets a chance to do that.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>