reafactor: Fix issues mentioned in code review
Change mentioning of clients to Represented persons or representee
to clarify what is being mention ( client is a concept in OAuth and
should be reserved).
Limited queries of delegation scope to just the name attribute since
that is the only one used.
Removed redundant AND keyword form query ( AND is implicit)
Removed Pick utility type where unneeded.