fix(j-s): add no cache fetchPolicy to UpdateCaseMutation
We are not using the apollo cache so we don't need to write to the cache
after mutation. Also, there is a problem with the test when writing to
the cache. We are mocking this request and Apollo insist of having all
the fields provided for Case, even if they are optional.