feat(service-portal): move feature flags to module definition (#8534)
* feat(service-portal): move feature flags to module definition

This gives us a couple of benefits:

- Allows us to refer to specific Features enum values with TypeScript checking.
- All active features are in one place.
- You have more control over the name of the feature flag which enables a service portal module.
- When you are cleaning up a feature flag, TypeScript will tell you everywhere it's used. No string magic.
- More consistent with how other code is feature flagged (application system, API routes).

Why:

Sometimes we use the same feature flag to enable logic in the service portal, API (and potentially other systems). If the feature flag has "servicePortal" in the name, it sounds like it is only used in the service portal. Recently there was an incident where someone was cleaning up one of these feature flags in the service portal but forgot to check if it was used in other places. This took down the feature in production. With this PR we can have a broader name, and TypeScript is likely to yell if you don't remove the feature flag everywhere in the repo.

* Add subjectType to nestjs feature flag service.

* Add test for subjectType feature flag attribute

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>