feat(auth): Add an internal procuring scope
There is a need for an scope similar to the internal scope but for Procuring holders.
This would allow delegation holders access to subsystems like applications conditionally.
There has already been implemented a custom scope in one such instance, the no-debt-certificate scope.
This could replace that scope and used in new instances f.e. the financial statement Inao application.