feat(auth): Add an internal procuring scope (#8831)
* feat(auth): Add an internal procuring scope

There is a need for an scope similar to the internal scope but for Procuring holders.
This would allow delegation holders access to subsystems like applications conditionally.
There has already been implemented a custom scope in one such instance, the no-debt-certificate scope.
This could replace that scope and used in new instances f.e. the financial statement Inao application.

* feat: Add scope into seed data

* feat: Phase out noDebt scope

Remove requesting the noDebt scope in the application layer
NoDebt api accepts now the new scope as well as the one being phased out

* Update libs/auth-api-lib/seeders/data/scope-internal-procuring.ts

Co-authored-by: Eiríkur Heiðar Nilsson <eirikur@nilsson.is>

* Update libs/auth-api-lib/seeders/data/scope-internal-procuring.ts

Co-authored-by: Eiríkur Heiðar Nilsson <eirikur@nilsson.is>

Co-authored-by: Eiríkur Heiðar Nilsson <eirikur@nilsson.is>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>