fix(j-s): Remove .pdf ending from indictment case files
Set the user generated filename automatically when a file is created. In the future we might want to rename the field. The reason for setting it is so it can be used to display the name of the file without .pdf at the end, making it easier for prosecutors to arrange indictment case files, as they dont need to manually remove the .pdf ending of each one.