chore(infra): Don't prefix resources with web-
Currently we are deploying resources with the services' release name.
However, this is currently `web` everywhere, i.e. redundant.

This prefix is inconsistent since we only add the prefix in some cases,
e.g. the service label is `service:name-of-service`, while
`kube_service:web-name-of-service`. I hypothesize that this is causing
issues for DataDog's Automatic Service Tagger.

This PR aims to fix this mismatch and unify `<label>` and
`kube_<label>`.