fix(auth): Speed up delegation queries HOTFIX (#13866)
* fix(auth): Speed up delegation queries

These queries are super slow in production because they have a LEFT OUTER JOIN against all delegations in the system containing the respective api scopes.

The original goal is to filter api scopes based on custom delegations which the authenticated user has against the current subject. So this PR changes removes the offending joins and instead conditionally queries the list of custom delegation scopes before the main query, which can then match against them with a simple IN filter.

* Review feedback

---------

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>